pyproject.toml
[tool.poetry]
name = "beta-sierra"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
sse-starlette = "^1.6.5"
tomli-w = "^1.0.0"
uvicorn = "^0.23.2"
fastapi = "^0.110.0"
langserve = "^0.0.46"
langchain-cli = "*"
pirate-speak = { path = "packages/pirate-speak" }
python-dotenv = "^1.0.0"
langchain = "^0.1.10"
langchain-openai = "^0.1.3"
faiss-cpu = "^1.8.0"
langgraph = "^0.0.38"
bs4 = "^0.0.2"
chromadb = "^0.4.24"
langchain-cohere = "^0.1.3"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.23.2"
pygithub = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
